backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.9.0
pydantic-settings==2.5.0
python-dotenv==1.0.1
python-multipart==0.0.9
openai==1.50.0
elevenlabs==1.9.0
google-cloud-texttospeech==2.17.0
google-cloud-speech==2.27.0
google-auth==2.35.0
azure-storage-blob==12.22.0
httpx==0.27.0
pytest==8.3.0
pytest-asyncio==0.24.0
